+++ lib/components/Actions/Link/index.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
+import { ComponentProps } from "react"
 import { Link } from "."
 
 const meta = {
@@ -10,12 +11,23 @@ const meta = {
   tags: ["autodocs"],
   args: {
     children: "This is a link",
-    href: "https://www.youtube.com/watch?v=dQw4w9WgXcQ",
-    target: "_blank",
+    href: "/",
   },
-} satisfies Meta<typeof Link>
+} satisfies Meta<ComponentProps<typeof Link>>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
-export const Basic: Story = {}
+export const Default: Story = {}
+export const TargetBlank: Story = {
+  args: {
+    target: "_blank",
+    children: "This link opens in a new tab",
+  },
+}
+
+export const AsText: Story = {
+  args: {
+    variant: "text",
+  },
+}
