+++ lib/components/Actions/exports.tsx
@@ -1,1 +1,13 @@
+import { Component } from "@/lib/component"
+import { Link as LinkComponent } from "./Link"
+
 export * from "./Button"
+export type { LinkProps } from "./Link"
+
+export const Link = Component(
+  {
+    name: "Link",
+    type: "info",
+  },
+  LinkComponent
+)
