+++ lib/components/Utilities/exports.tsx
@@ -1,6 +1,5 @@
 import { Component } from "@/lib/component"
 import { Icon as IconComponent } from "./Icon"
-import { Link as LinkComponent } from "./Link"
 import { ScrollArea as ScrollAreaComponent } from "./ScrollArea"
 
 export const ScrollArea = Component(
@@ -17,12 +16,4 @@ export const Icon = Component(
     type: "info",
   },
   IconComponent
-)
-
-export const Link = Component(
-  {
-    name: "Link",
-    type: "info",
-  },
-  LinkComponent
 )
